+++ Tests/BoilerplateTests.pas
@@ -487,6 +487,15 @@ begin
     UnregisterCustomOptions('/index.html');
     WhenRequest('/index.html');
     ThenOutHeaderValueIs('Cache-Control', 'no-transform');
+
+    GivenClearServer;
+    GivenAssets;
+    GivenOptions([bpoSetCacheNoTransform]);
+    RegisterCustomOptions('/index.html', [bpoSetCacheNoCache]);
+    RegisterCustomOptions('/404.html', [bpoSetCacheNoCache]);
+    UnregisterCustomOptions(['/index.html', '/404.html']);
+    WhenRequest('/index.html');
+    ThenOutHeaderValueIs('Cache-Control', 'no-transform');
   end;
 end;
 
@@ -1067,6 +1076,29 @@ begin
     GivenOptions([bpoSetCacheNoTransform]);
     RegisterCustomOptions('/index.html', [bpoSetCacheNoCache]);
     WhenRequest('/index.html');
+    ThenOutHeaderValueIs('Cache-Control', 'no-cache');
+    ThenRequestResultIs(HTML_SUCCESS);
+
+    GivenClearServer;
+    GivenAssets;
+    GivenOptions([bpoSetCacheNoTransform]);
+    WhenRequest('/404.html');
+    ThenOutHeaderValueIs('Cache-Control', 'no-transform');
+    ThenRequestResultIs(HTML_SUCCESS);
+
+    GivenClearServer;
+    GivenAssets;
+    GivenOptions([bpoSetCacheNoTransform]);
+    RegisterCustomOptions('/index.html', [bpoSetCacheNoCache]);
+    WhenRequest('/index.html');
+    ThenOutHeaderValueIs('Cache-Control', 'no-cache');
+    ThenRequestResultIs(HTML_SUCCESS);
+
+    GivenClearServer;
+    GivenAssets;
+    GivenOptions([bpoSetCacheNoTransform]);
+    RegisterCustomOptions(['/index.html', '/404.html'], [bpoSetCacheNoCache]);
+    WhenRequest('/404.html');
     ThenOutHeaderValueIs('Cache-Control', 'no-cache');
     ThenRequestResultIs(HTML_SUCCESS);
   end;